refactor scanning methods to separate file
